# openapi-spec/api-spec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LineUp.Backend | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/schedule/{guid}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Schedule"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Schedule"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/schedule": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ScheduleDto": {
        "required": [
          "dateCoverage",
          "startTime",
          "endTime",
          "schedulePreferences"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "guid": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "dateCoverage": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "shiftAssignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShiftAssignment"
            }
          },
          "schedulePreferences": {
            "$ref": "#/components/schemas/SchedulePreferences"
          }
        }
      },
      "SchedulePreferences": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "minutesPerSlot": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ShiftAssignment": {
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
}